Split test code in `conform-zod.spec.ts` into separate schema
