+++ packages/conform-zod/tests/coercion/schema/bigint.spec.ts
@@ -1,0 +1,59 @@
+import { describe, test, expect } from 'vitest';
+import { coerceFormValue } from '../../../coercion';
+import { z } from 'zod';
+import { getResult } from '../../helpers/zod';
+
+describe('coercion', () => {
+	describe('z.bigint', () => {
+		test('should pass bigint', () => {
+			const schema = z
+				.bigint({
+					error: (ctx) => {
+						if (ctx.input === undefined) {
+							return 'required';
+						}
+						return 'invalid';
+					},
+				})
+				.min(1n, 'min')
+				.max(10n, 'max')
+				.multipleOf(2n, 'step');
+			const file = new File([], '');
+
+			expect(getResult(coerceFormValue(schema).safeParse(''))).toEqual({
+				success: false,
+				error: {
+					'': ['required'],
+				},
+			});
+			expect(getResult(coerceFormValue(schema).safeParse('abc'))).toEqual({
+				success: false,
+				error: {
+					'': ['invalid'],
+				},
+			});
+			expect(getResult(coerceFormValue(schema).safeParse(file))).toEqual({
+				success: false,
+				error: {
+					'': ['invalid'],
+				},
+			});
+			expect(getResult(coerceFormValue(schema).safeParse(' '))).toEqual({
+				success: false,
+				error: {
+					'': ['invalid'],
+				},
+			});
+			expect(getResult(coerceFormValue(schema).safeParse('5'))).toEqual({
+				success: false,
+				error: {
+					'': ['step'],
+				},
+			});
+			expect(getResult(coerceFormValue(schema).safeParse('4'))).toEqual({
+				success: true,
+				data: 4n,
+			});
+		});
+	});
+});
